// src/services/leads/post/types.ts
export interface LeadBody {
  dateAndTime: string;
  date: string;
  make: string;
  rto: string;
  year: string;
  model: string;
  fuelType: string;
  transmission: string;
  variant: string;
  ownershipNumber: string;
  odometerReading: string;
  sellingPlan: string;
  sellerMobileNumber: string;
  source: string;
  purpose: string;
  registrationNumber: string;
  proposeOfInspection: "SELLING";
}